+++ server/src/main/java/com/adobe/testing/s3mock/dto/Retention.java
@@ -22,6 +22,7 @@ import com.fasterxml.jackson.annotation.JsonRootName;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import java.time.Instant;
+import java.util.Objects;
 
 /**
  * <a href="https://docs.aws.amazon.com/AmazonS3/latest/API/API_control_S3Retention.html">API Reference</a>.
@@ -63,5 +64,23 @@ public class Retention {
 
   public void setRetainUntilDate(Instant retainUntilDate) {
     this.retainUntilDate = retainUntilDate;
+  }
+
+  @Override
+  public boolean equals(Object o) {
+    if (this == o) {
+      return true;
+    }
+    if (o == null || getClass() != o.getClass()) {
+      return false;
+    }
+    Retention retention = (Retention) o;
+    return mode == retention.mode && Objects.equals(retainUntilDate,
+        retention.retainUntilDate);
+  }
+
+  @Override
+  public int hashCode() {
+    return Objects.hash(mode, retainUntilDate);
   }
 }