+++ server/src/main/java/com/adobe/testing/s3mock/util/AwsHttpParameters.java
@@ -23,6 +23,8 @@ public class AwsHttpParameters {
 
   private static final String NOT = "!";
 
+  public static final String ACL = "acl";
+  public static final String NOT_ACL = NOT + ACL;
   public static final String CONTINUATION_TOKEN = "continuation-token";
   public static final String DELETE = "delete";
   public static final String ENCODING_TYPE = "encoding-type";