+++ server/src/test/java/com/adobe/testing/s3mock/BucketControllerTest.java
@@ -76,7 +76,7 @@ import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 @SpringBootTest(classes = {S3MockConfiguration.class})
 class BucketControllerTest {
 
-  private static final Owner TEST_OWNER = new Owner(123, "s3-mock-file-store");
+  private static final Owner TEST_OWNER = new Owner("123", "s3-mock-file-store");
   private static final ObjectMapper MAPPER = new XmlMapper();
   private static final String TEST_BUCKET_NAME = "test-bucket";
   private static final Bucket TEST_BUCKET =