+++ server/src/test/java/com/adobe/testing/s3mock/dto/ListBucketResultTest.java
@@ -42,7 +42,7 @@ class ListBucketResultTest {
       S3Object s3Object =
           new S3Object("key" + i, "2009-10-12T17:50:30.000Z",
               "\"fba9dede5f27731c9771645a39863328\"", "434234", StorageClass.STANDARD,
-              new Owner(10L + i, "displayName"));
+              new Owner(String.valueOf(10L + i), "displayName"));
       s3ObjectList.add(s3Object);
     }
     return s3ObjectList;